# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CurificBackend Copy 2",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://3.82.115.123:5001/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/user/appointment/get": {
			"get": {
				"tags": ["Appointment"],
				"summary": "Get customer Appointment",
				"operationId": "GetcustomerAppointment",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/appointment/get": {
			"get": {
				"tags": ["Appointment"],
				"summary": "Get Appointment",
				"operationId": "GetAppointment",
				"parameters": [],
				"responses": {
					"default": {
						"description": "",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m20042"
								},
								"example": {
									"success": true,
									"message": "Appointment get successfully",
									"data": [
										{
											"symptoms": ["62b046d2d5debb823497dc1c"],
											"isActive": true,
											"drId": null,
											"consultModel": "VIDEO",
											"prescription": null,
											"notes": null,
											"deletedAt": null,
											"deletedBy": null,
											"isDeleted": false,
											"_id": "62d65ebdf511bf5509c3b62c",
											"date": "2022-07-19T02:30:00.000Z",
											"startTime": "8:00 AM",
											"endTime": "8:30 AM",
											"patientId": "62d51064f411fb5f8a3c4716",
											"price": 700,
											"createdAt": "2022-07-19T07:35:25.733Z",
											"updatedAt": "2022-07-19T07:35:25.733Z",
											"__v": 0
										},
										{
											"symptoms": ["62b046d2d5debb823497dc1c"],
											"isActive": true,
											"drId": null,
											"consultModel": "VIDEO",
											"prescription": null,
											"notes": null,
											"deletedAt": null,
											"deletedBy": null,
											"isDeleted": false,
											"_id": "62d668a1f511bf5509c3b632",
											"date": "2023-07-19T18:30:00.000Z",
											"startTime": "8:00 A",
											"endTime": "8:30 A",
											"patientId": "62d51064f411fb5f8a3c4716",
											"price": 700,
											"createdAt": "2022-07-19T08:17:37.343Z",
											"updatedAt": "2022-07-19T08:18:14.205Z",
											"__v": 0
										}
									],
									"totalCount": 2
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/appointment/update/{id}": {
			"put": {
				"tags": ["Appointment"],
				"summary": "Appointment Update",
				"operationId": "AppointmentUpdate",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AppointmentUpdateRequest"
							},
							"example": {
								"date": "08-18-2022",
								"startTime": "8:00 A",
								"endTime": "8:30 A",
								"pain": ["62b046d2d5debb823497dc1c"]
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m20043"
								},
								"example": {
									"date": "08-18-2022",
									"startTime": "8:00 A",
									"endTime": "8:30 A",
									"pain": ["62b046d2d5debb823497dc1c"]
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
		"/chat-room/sendMessage": {
			"post": {
				"tags": ["Chat Room"],
				"summary": "send Message",
				"operationId": "sendMessage",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/sendMessageRequest"
							},
							"example": {
								"senderId": "62ef944d81cbda2d545341ae",
								"content": "sadsad",
								"roomId": "62fcc7311159157e02bdb4ac"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m5004"
								},
								"example": {
									"success": true,
									"message": "Message sent successfully"
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"m5004": {
				"title": "m5004",
				"required": ["success", "message"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"success": false,
					"message": "Device Already available"
				}
			},
...
			"Data17": {
				"title": "Data17",
				"required": [
					"symptoms",
					"isActive",
					"drId",
					"consultModel",
					"prescription",
					"notes",
					"deletedAt",
					"deletedBy",
					"isDeleted",
					"_id",
					"date",
					"startTime",
					"endTime",
					"patientId",
					"price",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"properties": {
					"symptoms": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					},
					"isActive": {
						"type": "boolean"
					},
					"drId": {
						"type": "string",
						"nullable": true
					},
					"consultModel": {
						"type": "string"
					},
					"prescription": {
						"type": "string",
						"nullable": true
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"deletedAt": {
						"type": "string",
						"nullable": true
					},
					"deletedBy": {
						"type": "string",
						"nullable": true
					},
					"isDeleted": {
						"type": "boolean"
					},
					"_id": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"patientId": {
						"type": "string"
					},
					"price": {
						"type": "integer",
						"format": "int32"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"symptoms": ["62b046d2d5debb823497dc1c"],
					"isActive": true,
					"drId": null,
					"consultModel": "VIDEO",
					"prescription": null,
					"notes": null,
					"deletedAt": null,
					"deletedBy": null,
					"isDeleted": false,
					"_id": "62d6749958eb62639e632ffd",
					"date": "2022-07-18T02:30:00.000Z",
					"startTime": "8:00 A",
					"endTime": "8:30 A",
					"patientId": "62d51064f411fb5f8a3c4716",
					"price": 700,
					"createdAt": "2022-07-19T09:08:41.218Z",
					"updatedAt": "2022-07-19T09:08:41.218Z",
					"__v": 0
				}
			},
			"m20042": {
				"title": "m20042",
				"required": ["success", "message", "data", "totalCount"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Data17"
						},
						"description": ""
					},
					"totalCount": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"success": true,
					"message": "Appointment get successfully",
					"data": [
						{
							"symptoms": ["62b046d2d5debb823497dc1c"],
							"isActive": true,
							"drId": null,
							"consultModel": "VIDEO",
							"prescription": null,
							"notes": null,
							"deletedAt": null,
							"deletedBy": null,
							"isDeleted": false,
							"_id": "62d65ebdf511bf5509c3b62c",
							"date": "2022-07-19T02:30:00.000Z",
							"startTime": "8:00 AM",
							"endTime": "8:30 AM",
							"patientId": "62d51064f411fb5f8a3c4716",
							"price": 700,
							"createdAt": "2022-07-19T07:35:25.733Z",
							"updatedAt": "2022-07-19T07:35:25.733Z",
							"__v": 0
						},
						{
							"symptoms": ["62b046d2d5debb823497dc1c"],
							"isActive": true,
							"drId": null,
							"consultModel": "VIDEO",
							"prescription": null,
							"notes": null,
							"deletedAt": null,
							"deletedBy": null,
							"isDeleted": false,
							"_id": "62d668a1f511bf5509c3b632",
							"date": "2023-07-19T18:30:00.000Z",
							"startTime": "8:00 A",
							"endTime": "8:30 A",
							"patientId": "62d51064f411fb5f8a3c4716",
							"price": 700,
							"createdAt": "2022-07-19T08:17:37.343Z",
							"updatedAt": "2022-07-19T08:18:14.205Z",
							"__v": 0
						}
					],
					"totalCount": 2
				}
			},
			"AppointmentUpdateRequest": {
				"title": "AppointmentUpdateRequest",
				"required": ["date", "startTime", "endTime", "pain"],
				"type": "object",
				"properties": {
					"date": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"pain": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"date": "08-18-2022",
					"startTime": "8:00 A",
					"endTime": "8:30 A",
					"pain": ["62b046d2d5debb823497dc1c"]
				}
			},
			"m20043": {
				"title": "m20043",
				"required": ["date", "startTime", "endTime", "pain"],
				"type": "object",
				"properties": {
					"date": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"pain": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"date": "08-18-2022",
					"startTime": "8:00 A",
					"endTime": "8:30 A",
					"pain": ["62b046d2d5debb823497dc1c"]
				}
			},
...
			"sendMessageRequest": {
				"title": "sendMessageRequest",
				"required": ["senderId", "content", "roomId"],
				"type": "object",
				"properties": {
					"senderId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"roomId": {
						"type": "string"
					}
				},
				"example": {
					"senderId": "62ef944d81cbda2d545341ae",
					"content": "sadsad",
					"roomId": "62fcc7311159157e02bdb4ac"
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}